Moved SSTableFormat to sstable package and fixed PR issues
