+++ internal/sstable/builder.go
@@ -3,7 +3,6 @@ package sstable
 import (
 	"bytes"
 	"encoding/binary"
-	"fmt"
 	"github.com/gammazero/deque"
 	"github.com/samber/mo"
 	"github.com/slatedb/slatedb-go/gen"
@@ -22,6 +21,8 @@ type Table struct {
 	Bloom mo.Option[bloom.Filter]
 
 	// Blocks is a list of blocks contained in the Table
+	// NOTE: The final block added to the queue includes
+	// the encoded form of sstable.Index and sstable.Info
 	Blocks *deque.Deque[[]byte]
 }
 
@@ -36,7 +37,7 @@ type Table struct {
 // |  |  +-----------------------------------+  |  |
 // |  |  |  block.Block                      |  |  |
 // |  |  |  +-------------------------------+|  |  |
-// |  |  |  |  List of types.KeyValue pairs  |  |  |
+// |  |  |  |  List of KeyValue pairs        |  |  |
 // |  |  |  |  +---------------------------+ |  |  |
 // |  |  |  |  |  Key Length (2 bytes)     | |  |  |
 // |  |  |  |  |  Key                      | |  |  |
@@ -60,23 +61,25 @@ type Table struct {
 // |  +-----------------------------------------+  |
 // |                                               |
 // |  +-----------------------------------------+  |
-// |  |  sstable.Index                          |  |
+// |  |  flatbuf.SsTableIndexT                  |  |
 // |  |  (List of Block Offsets)                |  |
-// |  |  - Block Offset (End of Block)          |  |
+// |  |  - Block Offset (Start of Block)        |  |
 // |  |  - FirstKey of this Block               |  |
 // |  |  ...                                    |  |
 // |  +-----------------------------------------+  |
 // |                                               |
 // |  +-----------------------------------------+  |
-// |  |  sstable.Info                           |  |
-// |  |  - Offset of BloomFilter                |  |
-// |  |  - Length of BloomFilter                |  |
-// |  |  - Offset of sstable.Index              |  |
-// |  |  - Length of sstable.Index              |  |
+// |  |  flatbuf.SsTableInfoT                   |  |
+// |  |  - FirstKey of the SSTable              |  |
+// |  |  - Offset of bloom.Filter               |  |
+// |  |  - Length of bloom.Filter               |  |
+// |  |  - Offset of flatbuf.SsTableIndexT      |  |
+// |  |  - Length of flatbuf.SsTableIndexT      |  |
+// |  |  - The Compression Codec                |  |
 // |  +-----------------------------------------+  |
 // |                                               |
 // |  +-----------------------------------------+  |
-// |  |  Offset of sstable.Info (4 bytes)       |  |
+// |  |  Offset of SsTableInfoT (4 bytes)       |  |
 // |  +-----------------------------------------+  |
 // +-----------------------------------------------+
 type Builder struct {
@@ -115,14 +118,14 @@ type Builder struct {
 // Config specifies how SSTable is Encoded and Decoded
 type Config struct {
 	// BlockSize is the size of each block in the SSTable
-	BlockSize int
+	BlockSize uint64
 
 	// MinFilterKeys is the minimum number of keys that must exist in the SSTable
 	// before a bloom filter is created. Reads on SSTables with a small number
 	// of items is faster than looking up in a bloom filter.
-	MinFilterKeys int
+	MinFilterKeys uint32
 
-	FilterBitsPerKey int
+	FilterBitsPerKey uint32
 
 	// The codec used to compress new SSTables. The compression codec used in
 	// existing SSTables already written disk is encoded into the SSTableInfo and
@@ -256,7 +259,7 @@ func (b *Builder) Build() (*Table, error) {
 	buf = append(buf, compressedIndexBlock...)
 
 	metaOffset := b.currentLen + uint64(len(buf))
-	firstKey, _ := b.firstKey.Get()
+	firstKey, _ := b.sstFirstKey.Get()
 
 	sstInfo := &Info{
 		FirstKey:         bytes.Clone(firstKey),
@@ -279,46 +282,14 @@ func (b *Builder) Build() (*Table, error) {
 	}, nil
 }
 
-// PrettyPrint returns a string representation of the SSTable in a human-readable format
-func PrettyPrint(table *Table) string {
-	var buf bytes.Buffer
-
-	// Print SSTable Info
-	_, _ = fmt.Fprintf(&buf, "SSTable Info:\n")
-	_, _ = fmt.Fprintf(&buf, "  First Key: %s\n", string(table.Info.FirstKey))
-	_, _ = fmt.Fprintf(&buf, "  Index Offset: %d\n", table.Info.IndexOffset)
-	_, _ = fmt.Fprintf(&buf, "  Index Length: %d\n", table.Info.IndexLen)
-	_, _ = fmt.Fprintf(&buf, "  Filter Offset: %d\n", table.Info.FilterOffset)
-	_, _ = fmt.Fprintf(&buf, "  Filter Length: %d\n", table.Info.FilterLen)
-	_, _ = fmt.Fprintf(&buf, "  Compression Codec: %s\n", table.Info.CompressionCodec)
-
-	// Print Bloom Filter info if present
-	if filter, ok := table.Bloom.Get(); ok {
-		_, _ = fmt.Fprintf(&buf, "\nBloom Filter:\n")
-		_, _ = fmt.Fprintf(&buf, "  Number of Probes: %d\n", filter.NumProbes)
-		_, _ = fmt.Fprintf(&buf, "  Data Length: %d\n", len(filter.Data))
-	} else {
-		_, _ = fmt.Fprintf(&buf, "\nNo Bloom Filter\n")
-	}
-
-	// Print Blocks
-	_, _ = fmt.Fprintf(&buf, "\nBlocks:\n")
+// EncodeTable encodes the provided sstable.Table into the
+// SSTable format as []byte.
+func EncodeTable(table *Table) []byte {
+	var result []byte
 	for i := 0; i < table.Blocks.Len(); i++ {
-		blockData := table.Blocks.At(i)
-		_, _ = fmt.Fprintf(&buf, "  Block %d:\n", i)
-		_, _ = fmt.Fprintf(&buf, "    Length: %d bytes\n", len(blockData))
-
-		// Decode and print the block
-		//var decodedBlock block.Block
-		//if err := block.Decode(&decodedBlock, blockData); err == nil {
-		//	_, _ = fmt.Fprintf(&buf, "%s", block.PrettyPrint(&decodedBlock))
-		//} else {
-		//	_, _ = fmt.Fprintf(&buf, "  Error decoding block: %v\n", err)
-		//}
-		//_, _ = fmt.Fprintf(&buf, "\n")
+		result = append(result, table.Blocks.At(i)...)
 	}
-
-	return buf.String()
+	return result
 }
 
 // TODO(thrawn01): Rename this to sstable.decode which is only used by SSTableFormat
