+++ slatedb/table_store.go
@@ -29,14 +29,14 @@ import (
 type TableStore struct {
 	mu            sync.RWMutex
 	bucket        objstore.Bucket
-	sstFormat     *SSTableFormat
+	sstFormat     *sstable.SSTableFormat
 	rootPath      string
 	walPath       string
 	compactedPath string
 	filterCache   otter.Cache[sstable.ID, mo.Option[bloom.Filter]]
 }
 
-func NewTableStore(bucket objstore.Bucket, format *SSTableFormat, rootPath string) *TableStore {
+func NewTableStore(bucket objstore.Bucket, format *sstable.SSTableFormat, rootPath string) *TableStore {
 	cache, err := otter.MustBuilder[sstable.ID, mo.Option[bloom.Filter]](1000).Build()
 	common.AssertTrue(err == nil, "")
 	return &TableStore{
@@ -75,14 +75,14 @@ func (ts *TableStore) getWalSSTList(walIDLastCompacted uint64) ([]uint64, error)
 func (ts *TableStore) TableWriter(sstID sstable.ID) *EncodedSSTableWriter {
 	return &EncodedSSTableWriter{
 		sstID:         sstID,
-		builder:       ts.sstFormat.tableBuilder(),
+		builder:       ts.sstFormat.TableBuilder(),
 		tableStore:    ts,
 		blocksWritten: 0,
 	}
 }
 
 func (ts *TableStore) TableBuilder() *sstable.Builder {
-	return ts.sstFormat.tableBuilder()
+	return ts.sstFormat.TableBuilder()
 }
 
 func (ts *TableStore) WriteSST(id sstable.ID, encodedSST *sstable.Table) (*sstable.Handle, error) {
@@ -105,7 +105,7 @@ func (ts *TableStore) WriteSST(id sstable.ID, encodedSST *sstable.Table) (*sstab
 
 func (ts *TableStore) OpenSST(id sstable.ID) (*sstable.Handle, error) {
 	obj := ReadOnlyObject{ts.bucket, ts.sstPath(id)}
-	sstInfo, err := ts.sstFormat.readInfo(obj)
+	sstInfo, err := ts.sstFormat.ReadInfo(obj)
 	if err != nil {
 		logger.Error("unable to open table", zap.Error(err))
 		return nil, err
@@ -116,11 +116,11 @@ func (ts *TableStore) OpenSST(id sstable.ID) (*sstable.Handle, error) {
 
 func (ts *TableStore) ReadBlocks(sstHandle *sstable.Handle, blocksRange common.Range) ([]block.Block, error) {
 	obj := ReadOnlyObject{ts.bucket, ts.sstPath(sstHandle.Id)}
-	index, err := ts.sstFormat.readIndex(sstHandle.Info, obj)
+	index, err := ts.sstFormat.ReadIndex(sstHandle.Info, obj)
 	if err != nil {
 		return nil, err
 	}
-	return ts.sstFormat.readBlocks(sstHandle.Info, index, blocksRange, obj)
+	return ts.sstFormat.ReadBlocks(sstHandle.Info, index, blocksRange, obj)
 }
 
 // Reads specified blocks from an SSTable using the provided index.
@@ -130,7 +130,7 @@ func (ts *TableStore) ReadBlocksUsingIndex(
 	index *sstable.Index,
 ) ([]block.Block, error) {
 	obj := ReadOnlyObject{ts.bucket, ts.sstPath(sstHandle.Id)}
-	return ts.sstFormat.readBlocks(sstHandle.Info, index, blocksRange, obj)
+	return ts.sstFormat.ReadBlocks(sstHandle.Info, index, blocksRange, obj)
 }
 
 func (ts *TableStore) cacheFilter(sstID sstable.ID, filter mo.Option[bloom.Filter]) {
@@ -148,7 +148,7 @@ func (ts *TableStore) ReadFilter(sstHandle *sstable.Handle) (mo.Option[bloom.Fil
 	}
 
 	obj := ReadOnlyObject{ts.bucket, ts.sstPath(sstHandle.Id)}
-	filtr, err := ts.sstFormat.readFilter(sstHandle.Info, obj)
+	filtr, err := ts.sstFormat.ReadFilter(sstHandle.Info, obj)
 	if err != nil {
 		return mo.None[bloom.Filter](), err
 	}
@@ -159,7 +159,7 @@ func (ts *TableStore) ReadFilter(sstHandle *sstable.Handle) (mo.Option[bloom.Fil
 
 func (ts *TableStore) ReadIndex(sstHandle *sstable.Handle) (*sstable.Index, error) {
 	obj := ReadOnlyObject{ts.bucket, ts.sstPath(sstHandle.Id)}
-	index, err := ts.sstFormat.readIndex(sstHandle.Info, obj)
+	index, err := ts.sstFormat.ReadIndex(sstHandle.Info, obj)
 	if err != nil {
 		return nil, err
 	}
@@ -195,7 +195,7 @@ func (ts *TableStore) Clone() *TableStore {
 	return &TableStore{
 		mu:            sync.RWMutex{},
 		bucket:        ts.bucket,
-		sstFormat:     ts.sstFormat.clone(),
+		sstFormat:     ts.sstFormat.Clone(),
 		rootPath:      ts.rootPath,
 		walPath:       ts.walPath,
 		compactedPath: ts.compactedPath,
@@ -214,13 +214,13 @@ type EncodedSSTableWriter struct {
 	tableStore *TableStore
 
 	// TODO: we are using an unbounded slice of byte as buffer.
-	//  add a capacity for buffer and when buffer reaches the capacity
+	//  Add a capacity for buffer and when buffer reaches the capacity
 	//  it should be written to object storage
 	buffer        []byte
 	blocksWritten uint64
 }
 
-func (w *EncodedSSTableWriter) add(key []byte, value mo.Option[[]byte]) error {
+func (w *EncodedSSTableWriter) Add(key []byte, value mo.Option[[]byte]) error {
 	err := w.builder.Add(key, value)
 	if err != nil {
 		logger.Error("unable to Add key value", zap.Error(err))
@@ -239,7 +239,11 @@ func (w *EncodedSSTableWriter) add(key []byte, value mo.Option[[]byte]) error {
 	return nil
 }
 
-func (w *EncodedSSTableWriter) close() (*sstable.Handle, error) {
+func (w *EncodedSSTableWriter) Written() uint64 {
+	return w.blocksWritten
+}
+
+func (w *EncodedSSTableWriter) Close() (*sstable.Handle, error) {
 	encodedSST, err := w.builder.Build()
 	if err != nil {
 		logger.Error("unable to close SS Table", zap.Error(err))
